+++ src/client/src/config/tls.rs
@@ -5,11 +5,7 @@ use std::path::PathBuf;
 
 use tracing::info;
 use serde::{Deserialize, Serialize};
-use fluvio_future::net::{DomainConnector, DefaultTcpDomainConnector};
-use fluvio_future::native_tls::{
-    TlsDomainConnector, ConnectorBuilder, IdentityBuilder, X509PemBuilder, PrivateKeyBuilder,
-    CertBuilder, TlsAnonymousConnector,
-};
+use fluvio_future::net::{DomainConnector, DefaultDomainConnector};
 
 /// Describes whether or not to use TLS and how
 #[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
@@ -194,12 +190,17 @@ pub struct TlsPaths {
     pub ca_cert: PathBuf,
 }
 
+#[cfg(not(target_arch = "wasm32"))]
 impl TryFrom<TlsPolicy> for DomainConnector {
     type Error = IoError;
 
     fn try_from(config: TlsPolicy) -> Result<Self, Self::Error> {
+        use fluvio_future::native_tls::{
+            TlsDomainConnector, ConnectorBuilder, IdentityBuilder, X509PemBuilder,
+            PrivateKeyBuilder, CertBuilder, TlsAnonymousConnector,
+        };
         match config {
-            TlsPolicy::Disabled => Ok(Box::new(DefaultTcpDomainConnector::new())),
+            TlsPolicy::Disabled => Ok(Box::new(DefaultDomainConnector::new())),
             TlsPolicy::Anonymous => {
                 info!("Using anonymous TLS");
                 let connector: TlsAnonymousConnector = ConnectorBuilder::anonymous().build().into();
@@ -253,5 +254,14 @@ impl TryFrom<TlsPolicy> for DomainConnector {
                 )))
             }
         }
+    }
+}
+#[cfg(target_arch = "wasm32")]
+impl TryFrom<TlsPolicy> for DomainConnector {
+    type Error = IoError;
+
+    fn try_from(_config: TlsPolicy) -> Result<Self, Self::Error> {
+        info!("Using Default Domain connector for wasm");
+        Ok(Box::new(DefaultDomainConnector::new()))
     }
 }