+++ src/socket/src/sink.rs
@@ -36,7 +36,8 @@ impl FluvioSink {
     }
 
     pub fn id(&self) -> ConnectionFd {
-        self.fd
+        #[allow(clippy::clone_on_copy)]
+        self.fd.clone()
     }
 
     /// convert to shared sink
@@ -210,15 +211,17 @@ impl ExclusiveFlvSink {
     }
 
     pub fn id(&self) -> ConnectionFd {
-        self.fd
+        #[allow(clippy::clone_on_copy)]
+        self.fd.clone()
     }
 }
 
 impl Clone for ExclusiveFlvSink {
     fn clone(&self) -> Self {
+        #[allow(clippy::clone_on_copy)]
         Self {
             inner: self.inner.clone(),
-            fd: self.fd,
+            fd: self.fd.clone(),
         }
     }
 }