+++ src/spu/src/services/public/stream_fetch.rs
@@ -660,7 +660,7 @@ mod test {
 
     fn read_filter_from_path(filter_path: impl AsRef<Path>) -> Vec<u8> {
         let path = filter_path.as_ref();
-        std::fs::read(path).expect(&format!("Unable to read file {}", path.display()))
+        std::fs::read(path).unwrap_or_else(|_| panic!("Unable to read file {}", path.display()))
     }
 
     fn load_wasm_module(module_name: &str) -> Vec<u8> {