+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 * `spu_pool` to support wasm runtime. ([#1106](https://github.com/infinyon/fluvio/pull/1106))
 * Remove async trait for more wasm support to client ([#1108](https://github.com/infinyon/fluvio/pull/1108))
 * Better logging for SPU health check ([#1109](https://github.com/infinyon/fluvio/issues/1109))
+* fluvio-socket build for wasm32 ([#1111](https://github.com/infinyon/fluvio/issues/1111))
 
 ## Platform Version 0.8.2 - 2020-05-06
 * Fix Replication fail over with duplication ([#1052](https://github.com/infinyon/fluvio/pull/1052))